pyproject.toml
[project]
name = "qwen-local"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "pywin32>=306; sys_platform == 'win32'",
    "jieba>=0.42.0",
    "rank-bm25>=0.2.2",
    "bitsandbytes>=0.49.2",
    "fastapi>=0.115.0",
    "python-multipart>=0.0.9",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "python-docx>=1.0.0",
    "huggingface-hub>=1.6.0",
    "modelscope>=1.34.0",
    "python-dotenv>=1.2.2",
    "torch>=2.0.0",
    "transformers>=5.3.0",
    "uvicorn>=0.32.0",
]

# torch 由 setup.ps1 Step 4.5 单独安装 GPU 版
# 这里用官方源仅供 uv 做依赖解析，不影响实际下载
[tool.uv.sources]
torch = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[[tool.uv.index]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
explicit = true